+++ react-app-5/src/context/Books.js
@@ -1,0 +1,20 @@
+import { createContext, useState } from "react";
+const BooksContext = createContext();
+
+export function Provider({ children }) {
+  const [count, setCount] = useState(0);
+
+  const valueToShare = {
+    count,
+    incrementCount: () => {
+      setCount(count + 1);
+    },
+  };
+  return (
+    <BooksContext.Provider value={valueToShare}>
+      {children}
+    </BooksContext.Provider>
+  );
+}
+
+export default BooksContext;
